Fixed some compiler warnings

+++ src/game/collision.cpp
@@ -67,7 +67,7 @@ bool CCollision::IsTileSolid(int x, int y)
 int CCollision::IntersectLine(vec2 Pos0, vec2 Pos1, vec2 *pOutCollision, vec2 *pOutBeforeCollision)
 {
 	float d = distance(Pos0, Pos1);
-	int End(d+1);
+	int End((int)d+1);
 	vec2 Last = Pos0;
 	
 	for(int i = 0; i < End; i++)
